python/kubernetesscaler/requirements.txt
kubernetes==12.0.1
mysql-connector-python==8.0.23